Refactor transaction message codecs to group by transaction version
