+++ packages/transaction-messages/src/codecs/v0/address-table-lookup.ts
@@ -9,7 +9,7 @@ import {
 import { getArrayDecoder, getArrayEncoder, getStructDecoder, getStructEncoder } from '@solana/codecs-data-structures';
 import { getShortU16Decoder, getShortU16Encoder, getU8Decoder, getU8Encoder } from '@solana/codecs-numbers';
 
-import type { getCompiledAddressTableLookups } from '../compile/address-table-lookups';
+import type { getCompiledAddressTableLookups } from '../../compile/address-table-lookups';
 
 type AddressTableLookup = ReturnType<typeof getCompiledAddressTableLookups>[number];
